Rename notestore to object store, test indexing
